start position for grasping as with joint values
The start position of the Robotarm is set in joint values now.
Before this change it may happend that the elbow joint folded in the
direction of the table, in this case the robot can't grap the object in
a simple trajectory.
Now the elbow points always to the top so we don't get some complex
trajectorys.