Use of Gripper action server, remove of some sleeps
The gripper is now closed with the gripper action server, instead of
publishing the register states for the gripper.
Due to that the sleeps after the gripper motions could be removed, cause
the .move() funktion only returns when the motion is finished.
The new move_group for the gripper is named group_arm, the move_group
for the arm is changed to group_arm, so the different is clear.

The sleeps after adding and deleting objects to the scene stay, to get
sure that the model is realy the and will be respected by the
motionplanning.

The 10 seconds sleeping at the beginning when the arm moves to the start
position was set down to 5 seconds. 5 seconds should be enough for the
vision module to detect the object instead of the robotarm that moves
across the search area.