Final reimplementation of object recognition
For the object recognition we use the Pointcloud from the camera and
transform it into the world frame. In the next step we use CropBox to
set our region of interest. This region is about 50 cm abouve the table.
After that we cluster the points to get a pointcloud of the object.
For the x and y values we use the midpoint of this cloud. the z value is
the highest point and publish a transform with this values for the
object. In the next step we crate a Marker for the visualization in
rviz.

We exchange the old node in the launchfile with the new one.