Variable size of marker for object
Cloud will be transformed to the new link table_top now. The ROI and all
other necessary changes for this adjustment were made.
The size and position of the marker has a new calculation and depends on
the object height now.

The result is a cylinder which underneath is always on the surface of
the table and the height is the highest point of the object cloud and
therefor the height of the object.