Establish planning scene service
We use the new apply planning scene service instead of the publisher.
The attached objects are AttachedCollisionObjects now and the other
objects are normal CollisionObjects like before. There are two new
functions to remove the collision objects and attached collision
objects.
This has the advantage that we get rid of the 2sec sleeps after
attaching and detaching the can.

In this cause the code is cleaned up a bit.