Use of moveits pick interface
Due to the use of pick we get rid of some movements. The one above the
object, down to the object and closing the gripper.
Pick take care of the attachement and detachement of the object, in that
case we need to rewrite the removeObject function for the attached
objects.

Because of the last moveit updates, we don't need to add a collision
object to the scene after removing the attached object.